requirements.txt
Flask==3.0.0
Waitress==2.1.2
openpyxl==3.1.2
pandas==2.1.3
python-dotenv==1.0.0
supabase==2.3.4
